task/apollo-mutations: move all queries to /configs/queries.js
